user_service/requirements.txt
grpcio
grpcio-tools
pymongo
passlib
pyjwt